Add Ada.Real_Time "delay until" loop to test app.
The "surface" test app has an empty SDL polling busy loop that uses all
of a CPU core if allowed to. Using Ada.Real_Time's monotonic Clock,
along with "delay until", we can run the event loop at 60 Hz (as in the
modified example) or at any other rate, without being affected by any
system time changes.

The difference between the requested Ada.Real_Time.Time and the clock at
the start of the next loop is on the order of 0.150 ms on my PC.