Add 144 Hz animation loop to stream2 test app.
The "stream2" test app was changing the moose animation frame on every
cycle of the polling loop, with no delays in between frames. Changed the
code similarly to the previous commit for a different test app, but this
time use a 144 Hz (6_944 us) delay instead of 60 Hz, and update the
animation frame every 4 event loop updates, if either the left or right
cursor keys are currently pressed (but not both).